requirements.txt
flask
google-adk
google-genai
pandas
gunicorn
python-dotenv
